Adds a flash policy server.
The Flash fallback doesn't work when using TLS because Flash sends
a request for it's cross domain policy in the clear over a socket that's
supposed to be encrypted.
However, Flash also checks the port 843 for a policy file. This commit
adds an optionnal command line parameter to start a simple TCP server serving
the policy file.